fix(ci): extract heatmap toggles to satisfy fallow health gate
Move review-gated heatmap switches into HeatmapDisplayToggle so
SettingsDialog and Sidebar stay within the complexity baseline.
